move sensor dict to base material
