+++ src/fenicsxconcrete/finite_element_problem/base_material.py
@@ -8,7 +8,7 @@ from fenicsxconcrete.experimental_setup.cantilever_beam import CantileverBeam
 from fenicsxconcrete.experimental_setup.compression_cylinder import CompressionCylinder
 from fenicsxconcrete.experimental_setup.tensile_beam import TensileBeam
 from fenicsxconcrete.helper import LogMixin, Parameters
-from fenicsxconcrete.sensor_definition.base_sensor import Sensor, Sensors
+from fenicsxconcrete.sensor_definition.base_sensor import Sensor
 from fenicsxconcrete.unit_registry import ureg
 
 
@@ -49,7 +49,7 @@ class MaterialProblem(ABC, LogMixin):
         self.p = self.parameters.to_magnitude()
         self.experiment.p = self.p  # update experimental parameter list for use in e.g. boundary definition
 
-        self.sensors = Sensors()  # list to hold attached sensors
+        self.sensors = self.SensorDict()  # list to hold attached sensors
 
         # settin gup path for paraview output
         if not pv_path:
@@ -100,4 +100,30 @@ class MaterialProblem(ABC, LogMixin):
 
     def delete_sensor(self) -> None:
         del self.sensors
-        self.sensors = Sensors()
+        self.sensors = self.SensorDict()
+
+    class SensorDict(dict):
+        """
+        Dict that also allows to access the parameter p["parameter"] via the matching attribute p.parameter
+        to make access shorter
+
+        When to sensors with the same name are defined, the next one gets a number added to the name
+        """
+
+        def __getattr__(self, key):
+            return self[key]
+
+        def __setattr__(self, key, value):
+            assert key in self
+            self[key] = value
+
+        def __setitem__(self, initial_key: str, value: Sensor) -> None:
+            # check if key exists, if so, add a number to the name
+            i = 2
+            key = initial_key
+            if key in self:
+                while key in self:
+                    key = initial_key + str(i)
+                    i += 1
+
+            super().__setitem__(key, value)
